scripts/gdb: find vmlinux where it was before
Patch series "gdb script for kconfig and timer list".

This is a handful of changes to the kernel's gdb scripts to do some more
debugging with kgdb.  The first patch allows the vmlinux to be reloaded
from where it was specified on the command line so that this set of
scripts can be used from anywhere.  The second patch adds a script to dump
the config.gz to a file on the host debugging machine.  The third patch
adds some rb tree utilities and the last patch uses those rb tree walking
utilities to dump out the contents of /proc/timer_list from a system under
debug.

This patch (of 4):

If I run 'gdb <path/to/vmlinux>' and there's the vmlinux-gdb.py file there
I can properly see symbols and use the lx commands provided by the GDB
scripts.  But once I run 'lx-symbols' at the command prompt, gdb reloads
the vmlinux symbols assuming that this script was run from the directory
that has vmlinux at the root.  That isn't always true, but we could just
look and see what symbols were already loaded and use that instead.  Let's
do that so this can work by being invoked anywhere.

Link: http://lkml.kernel.org/r/20190325184522.260535-2-swboyd@chromium.org
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Cc: Douglas Anderson <dianders@chromium.org>
Cc: Nikolay Borisov <n.borisov.lkml@gmail.com>
Cc: Kieran Bingham <kbingham@kernel.org>
Cc: Jan Kiszka <jan.kiszka@siemens.com>
Cc: Jackie Liu <liuyun01@kylinos.cn>
Cc: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>