slob: only use list functions when safe to do so
Currently we call (indirectly) list_del() then we manually try to combat
the fact that the list may be in an undefined state by getting 'prev' and
'next' pointers in a somewhat contrived manner.  It is hard to verify that
this works for all initial states of the list.  Clearly the author (me)
got it wrong the first time because the 0day kernel testing robot managed
to crash the kernel thanks to this code.

All this is done in order to do an optimisation aimed at preventing
fragmentation at the start of a slab.  We can just skip this optimisation
any time the list is put into an undefined state since this only occurs
when an allocation completely fills the slab and in this case the
optimisation is unnecessary since we have not fragmented the slab by this
allocation.

Change the page pointer passed to slob_alloc_page() to be a double pointer
so that we can set it to NULL to indicate that the page was removed from
the list.  Skip the optimisation if the page was removed.

Found thanks to the kernel test robot, email subject:

	340d3d6178 ("mm/slob.c: respect list_head abstraction layer"):  kernel BUG at lib/list_debug.c:31!

Link: http://lkml.kernel.org/r/20190402032957.26249-2-tobin@kernel.org
Signed-off-by: Tobin C. Harding <tobin@kernel.org>
Reported-by: kernel test robot <lkp@intel.com>
Cc: Roman Gushchin <guro@fb.com>
Cc: Christoph Lameter <cl@linux.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: Stephen Rothwell <sfr@canb.auug.org.au>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>