mm: Introduce find_vma_rcu()
While under the RCU protection, a reference is taken on the VMA, so the
caller must call put_vma() once it not more needs the VMA structure.

When the VMA is about to be released, this is not done immediately but at
the end of the RCU grace period through vm_rcu_put(). This ensures that the
VMA remains allocated until the end the RCU grace period.

Since the vm_file pointer, if valid, is released in put_vma(), there is no
guarantee that the file pointer will be valid on the returned VMA.

Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>