include/linux/uaccess.h: add probe_user_read()
In powerpc code, there are several places implementing safe access to user
data.  This is sometimes implemented using probe_kernel_address() with
additional access_ok() verification, sometimes with get_user() enclosed in
a pagefault_disable()/enable() pair, etc:

    show_user_instructions()
    bad_stack_expansion()
    p9_hmi_special_emu()
    fsl_pci_mcheck_exception()
    read_user_stack_64()
    read_user_stack_32() on PPC64
    read_user_stack_32() on PPC32
    power_pmu_bhrb_to()

In the same spirit as probe_kernel_read(), this patch adds
probe_user_read().

probe_user_read() does the same as probe_kernel_read() but first checks
that it is really a user address.

The patch defines this function as a static inline so the "size" variable
can be examined for const-ness by the check_object_size() in
__copy_from_user_inatomic()

Link: http://lkml.kernel.org/r/39fb6c5a191025378676492e140dc012915ecaeb.1547652372.git.christophe.leroy@c-s.fr
Signed-off-by: Christophe Leroy <christophe.leroy@c-s.fr>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Acked-by: Murilo Opsfelder Araujo <muriloo@linux.ibm.com>
Cc: Kees Cook <keescook@chromium.org>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: Mike Rapoport <rppt@linux.ibm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>